requirements.txt
psycopg-binary==3.2.6
PyYAML==6.0.2
scapy==2.6.1
